requirements.txt
absl-py==1.3.0
aiosignal==1.2.0
astor==0.8.1
attrs==22.1.0
cached-property==1.5.2
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
distlib==0.3.6
filelock==3.4.1
frozenlist==1.2.0
gast==0.2.2
gitdb==4.0.9
GitPython==3.1.18
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.4
igraph==0.9.11
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
msgpack==1.0.4
numpy==1.19.5
opt-einsum==3.3.0
pandas==1.1.5
patsy==0.5.3
Pillow==8.4.0
pkg_resources==0.0.0
platformdirs==2.4.0
protobuf==3.19.6
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
ray==2.0.1
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
statsmodels==0.12.2
tabulate==0.8.10
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==3.1.0
torch==1.10.2
typing_extensions==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
